seata-solon-plugin: Adjust the configuration loading order
