+++ integration/solon-plugin/src/main/java/org/apache/seata/solon/autoconfigure/properties/core/ThreadFactoryProperties.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.solon.autoconfigure.properties;
+package org.apache.seata.solon.autoconfigure.properties.core;
 
 import org.apache.seata.core.rpc.netty.NettyBaseConfig.WorkThreadMode;
 import org.noear.solon.annotation.Configuration;
